move registry views to a dedicated app (#212)
