+++ config/urls.py
@@ -10,7 +10,7 @@ from django.views.generic import TemplateView
 
 from projectroles.views import HomeView
 from containers.urls import websocket_urlpatterns
-from containers.views import KioscRegistryProxyView
+from registry.views import KioscRegistryProxyView
 
 urlpatterns = (
     [
@@ -62,6 +62,8 @@ urlpatterns = (
         path('containers/', include('containers.urls')),
         # Containertemplates URLs
         path('containertemplates/', include('containertemplates.urls')),
+        # Reigstry URLs
+        path('registry', include('registry.urls')),
         # Docker registry proxy view (at this time, Docker clients do not support subpaths in a registry)
         re_path(
             r'^v2/(?P<path>.*)$',
