+++ containers/statemachines.py
@@ -1,7 +1,6 @@
 from asgiref.sync import async_to_sync
 from channels.layers import get_channel_layer
 
-import socket
 import shlex
 from urllib.parse import urlsplit
