+++ registry/tests/test_registry.py
